Address Spacy `None` issue (#22)
## Describe your changes
This version still works for the shorter csv file as merged before. 

This version takes care of None types and empty strings by changing them
to strings.

For the 400 MB utterance.csv, the run time is pretty long, and the
5-year Roberts case dataset might help shorten the runtime. If you would
like to check if this works for all data, sending a data job to
Hadoop/GCP might be worthwile.

## Checklist before requesting a review
<!--- These are suggested things you could add, but what you add will be
dependent on your repository's standards. --->
- [x] The code runs successfully.
- [x] I ran `make lint` and have made the bulk of changes that it has
requested.
- [x] The `Lint` GitHub Action step is passing.
  - Manually run `make format` to correct the errors.

---------

Co-authored-by: Jessup Jong <jessupjong@Jessups-MacBook-Pro-3.local>